// src/main/resources/angular/src/app/api/model/guildMember.ts
/**
 * Dum API
 * This is the DUM API
 *
 * OpenAPI spec version: 1.0.0
 * Contact: rbarbado@nexusgeographics.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { DocumentGuildMember } from './documentGuildMember';


export interface GuildMember { 
    memberId: number;
    memberName: string;
    dni: string;
    companyName: string;
    agremiatType: number;
    securityCode: string;
    zipCode: number;
    cityId: number;
    telephone: number;
    email: string;
    address: string;
    creationDate?: string;
    DocumentGuildMember?: DocumentGuildMember;
    guildId: number;
    deletedDate:string;
}